Kill $wgEnablePopups
It's not a feature flag because it disables whole extension as opposed to a
particular feature. It would result in weird situations where the extension is
present in Special:Version but does not work for no obvious reason. If an
extension needs to be disabled on WMF, its registration file should not be included,
as opposed to "feature flags". This, however, does not mean that I'm opposed to
real feature flags:)

Change-Id: I09bae50241057f727881f33d6f97f69cec7c0a10