Drop beta feature step testing in browser tests
Page previews is enabled by default on the beta cluster and
in master.
In production config we enable it for all wikis but enwiki
and dewiki.

We should thus stop testing this to bring our testing in line with
reality.

This allows our browser tests to run against the beta cluster as
well as per commit as well as reduces the amount of code we need to
port to Node.js

Additional changes:
* Remove redundant support pages and properties

Bug: T177262
Change-Id: Ia85e7b60774232ced35641f2671462ca0ef64f18